Extracting routing code from `NotificationDetails` into a `Router` struct.
